update request cards with new db schema
